simplify the cli implementation
